remove center
